* Setup basic Kivy GUI * main.py now contains the kivy interface * The file with processing functions has been renamed to process.py
